refactor: separated logics for each functionalities with FastAPI frameworks.
