+++ app/api/v1/schemas.py
@@ -1,6 +1,15 @@
 from pydantic import BaseModel
 
 
+class RootResponseBase(BaseModel):
+    pass
+
+
+class RootResponse(RootResponseBase):
+    path: str
+    detail: str
+
+
 class TranslateReqBody(BaseModel):
     country: str
